requirements.txt
Django>=2.0.2
django-auth-ldap>=1.3.0
django-formtools>=2.1
django-tables2>=1.21.1
psycopg2>=2.7.4
